feat: "post like기능 toggle형식으로 변경"
like 존재 시 삭제, 없을 시 추가

+++ src/main/java/com/example/InstagramCloneCoding/domain/postlike/application/PostLikeService.java
@@ -11,8 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.example.InstagramCloneCoding.domain.post.error.PostErrorCode.POST_NOT_FOUND;
-import static com.example.InstagramCloneCoding.domain.postlike.error.PostLikeError.POST_LIKE_EXIST;
-import static com.example.InstagramCloneCoding.domain.postlike.error.PostLikeError.POST_LIKE_NOT_FIND;
 
 @Service
 @Transactional
@@ -22,23 +20,16 @@ public class PostLikeService {
     private final PostLikeRepository postLikeRepository;
     private final PostRepository postRepository;
 
-    public void likePost(Member member, int postId) {
+    public void togglePostLike(Member member, int postId){
         Post post = postRepository.findById(postId).orElseThrow(() -> new RestApiException(POST_NOT_FOUND));
-        // 이미 좋아요 눌렀는 지 검사
-        postLikeRepository.findByMemberAndPost(member, post).ifPresent(postLike -> {
-            throw new RestApiException(POST_LIKE_EXIST);
-        });
+        PostLike postLike = postLikeRepository.findByMemberAndPost(member, post).orElse(null);
 
-        PostLike postLike = new PostLike(member, post);
-        postLikeRepository.save(postLike);
-    }
-
-    public void unlikePost(Member member, int postId) {
-        Post post = postRepository.findById(postId).orElseThrow(() -> new RestApiException(POST_NOT_FOUND));
-        // 좋아요 눌러있는 지 검사
-        PostLike postLike = postLikeRepository.findByMemberAndPost(member, post)
-                .orElseThrow(() -> new RestApiException(POST_LIKE_NOT_FIND));
-
-        postLikeRepository.delete(postLike);
+        // 이미 좋아요 누른 상태 -> 삭제
+        if(postLike!=null){
+            postLikeRepository.delete(postLike);
+            return;
+        }
+        // 좋아요 안누른 상태 -> 추가
+        postLikeRepository.save(new PostLike(member, post));
     }
 }
